backend/requirements.txt
Flask
flask-socketio
Flask-Cors
ecdsa
cryptography
pyjwt
